Extract Menu into a component
